refactor: Move edmx spec to data-model & formatter stuff to project
