+++ packages/odata2model/src/project/formatter/BaseFormatter.ts
