+++ packages/odata2model/src/project/formatter/NoopFormatter.ts
