+++ packages/odata2model/src/project/formatter/PrettierFormatter.ts
