+++ packages/odata2model/test/data-model/EntitySetDigestion.test.ts
@@ -2,7 +2,7 @@ import { ODataModelBuilder, ODataVersion } from "./builder/ODataModelBuilder";
 import { digest } from "../../src/data-model/DataModelDigestion";
 import { EmitModes, Modes, RunOptions } from "../../src/OptionModel";
 import { ModelTypes } from "../../src/data-model/DataTypeModel";
-import { OdataTypes } from "../../src/odata/ODataEdmxModel";
+import { OdataTypes } from "../../src/data-model/edmx/ODataEdmxModel";
 
 const NOOP_FN = () => {};
